chore: removed console logs from module review
